move getRepositories to common
